+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 LATEST CHANGES
 ==============
 
+2024-03-07
+----------
+- Add new parameters `gnss_fusion.h_covariance_mul` and `gnss_fusion.v_covariance_mul` to control the effects of the GNSS covariance
+- Add 
+
 2024-03-05
 ----------
 - Add ROI diagnostic